just get it compiling to start with:
audioQueue needs to be a var so we can set it in createAudioQueue.
Since we're going to set it, no need to create a ref in initialization,
just set it to nil to start.

for now, pass a nil pointer as unsafeDescription (fix next)

Create a local no-op output callback to pass to AudioQueueNewOutput

'data' is packet data, which is probably not what we want for the
userData argument of AudioQueueNewOutput.  I did the same thing as in
the other callbacks where I passed self and then unpacked it in the
callback.

I also removed some typealiases that were unnecessary because they're
already defined in the Audio library.